Sentinel API
The following commit introduces a sentinel and status API. Currently
the API only accepts `application/json` but in the future it would
be easy to add gRPC as a way for machine to machine communication.

This is the beginnings of this, there isn't much to look atm. More
work will be required before anything is useful.